Sadly i had to change most components and combine code. THis is due to react not being able to pass values back to parents from nested children without external libraries
