--- client/src/components/Topics/TopicTitle.js
@@ -1,16 +1,0 @@
-import React from "react";
-import Segment from "../Segment/Segment";
-
-const Titles = () => {
-  return (
-    <div>
-      <Segment title="Games" />
-      <Segment title="Movies" />
-      <Segment title="Props" />
-      <Segment title="Books" />
-      <Segment title="Community" />
-    </div>
-  );
-};
-
-export default Titles;
